+++ include/GafferSceneUI/LightTool.h
@@ -65,11 +65,6 @@ class GAFFERSCENEUI_API LightTool : public GafferSceneUI::SelectionTool
 
 		GAFFER_NODE_DECLARE_TYPE( GafferSceneUI::LightTool, LightToolTypeId, SelectionTool );
 
-		const IECore::PathMatcher selection() const;
-
-		using SelectionChangedSignal = Gaffer::Signals::Signal<void (LightTool &)>;
-		SelectionChangedSignal &selectionChangedSignal();
-
 	private :
 
 		GafferScene::ScenePlug *scenePlug();
@@ -95,8 +90,6 @@ class GAFFERSCENEUI_API LightTool : public GafferSceneUI::SelectionTool
 		bool m_handleTransformsDirty;
 
 		bool m_priorityPathsDirty;
-
-		SelectionChangedSignal m_selectionChangedSignal;
 
 		bool m_dragging;